feat: Multiple enhancements and fixes for Zmall
- Integrated Dashen MPGS
- Added TeleBirr InApp support for Android
- Updated promotional items location to filter promotions based on location
- Integrated AliExpress for both local and global screens
- Added AliExpress cart for both local and zmall global
- Fixed ETA time calculation issue with static times in checkout_screen.dart to resolve ETA server response issue
- Added show password feature on the login screen
- Modified UI for the store screen
- Implemented refresh indicator in home, store, and product screens
- Displayed AliExpress store name and navigation in the cart screen
- Implemented vehicleId and device_type for Kifiya screen  to specify payment methods
- Fixed horizontal overflow error in store cards on the home screen
- Improved banner image loading and provider in the home screen
- Enhanced automatic services and categories: Prediction, Events, Courier, AliExpress in HomeScreen
- Modified promotional items API to support location updates
- Updated deprecated packages and refactored deprecated code
- Ensured the Global screen always uses Ethiopia's settings/API's in login_screen.dart

+++ lib/borsa/borsa_screen.dart
@@ -71,7 +71,8 @@ class _BorsaScreenState extends State<BorsaScreen> {
         isLoading = false;
       });
     } else {
-      if (responseData['error_code'] == 999) {
+      if (responseData['error_code'] != null &&
+          responseData['error_code'] == 999) {
         await Service.saveBool('logged', false);
         await Service.remove('user');
         Navigator.pushReplacementNamed(context, LoginScreen.routeName);
@@ -79,7 +80,8 @@ class _BorsaScreenState extends State<BorsaScreen> {
       setState(() {
         isLoading = false;
       });
-      if (errorCodes['${responseData['error_code']}'] != null) {
+      if (responseData['error_code'] != null &&
+          errorCodes['${responseData['error_code']}'] != null) {
         ScaffoldMessenger.of(context).showSnackBar(SnackBar(
           content: Text("${errorCodes['${responseData['error_code']}']}"),
           backgroundColor: kSecondaryColor,
@@ -171,7 +173,7 @@ class _BorsaScreenState extends State<BorsaScreen> {
                       iconData: Icons.send,
                       title: "Transfer",
                       subtitle: "Send funds",
-                      color: kSecondaryColor.withOpacity(.8),
+                      color: kSecondaryColor.withValues(alpha: 0.8),
                       textColor: kPrimaryColor,
                       press: () {
                         showModalBottomSheet<void>(
